+++ src/backend/core/models.py
@@ -881,6 +881,7 @@ class Item(TreeModel, BaseModel):
             and self.type == ItemTypeChoices.FILE
             and self.upload_state == ItemUploadStateChoices.READY
         )
+        can_export = can_get and self.type == ItemTypeChoices.FOLDER
 
         return {
             "accesses_manage": can_manage,
@@ -891,6 +892,7 @@ class Item(TreeModel, BaseModel):
             "destroy": can_destroy,
             "download": can_get,
             "duplicate": can_duplicate,
+            "export": can_export,
             "hard_delete": can_hard_delete,
             "favorite": can_get and user.is_authenticated,
             "link_configuration": can_manage,